+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [0.5.3] UNRELEASED
 ### Added
 - Add `_fill_all_arguments` option to `filter_and_call`.
+- Add `parse_args_using_dataclass` function to parse program arguments using dataclass.
 
 ### Modified
 - Move CSV, JSON, JSONL and PICKLE functions to new `serialization` module.